% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETRep_Functions.R
\docType{data}
\name{colon3D}
\alias{colon3D}
\title{Data}
\format{
A list containing the information of an e-tube
}
\source{
Generated and stored in the package's `data/` folder.
}
\usage{
colon3D
}
\description{
A colon sample as an elliptical tube.
}
\keyword{datasets}
