% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETRep_Functions.R
\name{elliptical_Tube_Euclideanization}
\alias{elliptical_Tube_Euclideanization}
\title{Convert an ETRep to a Matrix in the Convex Transformed Space.}
\usage{
elliptical_Tube_Euclideanization(tube)
}
\arguments{
\item{tube}{A list containing the details of the ETRep.}
}
\value{
An n*6 matrix, where n is the number of spinal points, representing the ETRep in the transformed Euclidean convex space.
}
\description{
Convert an ETRep to a Matrix in the Convex Transformed Space.
}
\examples{
#Example
# Load tube
data("tube_A")
Euclideanized_Tube<- elliptical_Tube_Euclideanization(tube = tube_A)

}
