% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETRep_Functions.R
\docType{data}
\name{simulatedColons}
\alias{simulatedColons}
\title{Data}
\format{
Five simulated samples of elliptical tubes, modeled after a reference structure resembling a colon.
}
\source{
Generated and stored in the package's `data/` folder.
}
\usage{
simulatedColons
}
\description{
Simulated samples of e-tubes, modeled after a reference structure resembling a colon.
}
\keyword{datasets}
