% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETRep_Functions.R
\docType{data}
\name{tube_A}
\alias{tube_A}
\title{Data}
\format{
A list containing the information of an e-tube with 204 elliptical cross-sections
}
\source{
Generated and stored in the package's `data/` folder.
}
\usage{
tube_A
}
\description{
A tube with 204 elliptical cross-sections.
}
\keyword{datasets}
