% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETRep_Functions.R
\name{check_Tube_Legality}
\alias{check_Tube_Legality}
\title{Check the Legality of an Elliptical Tube (ETRep)}
\usage{
check_Tube_Legality(tube)
}
\arguments{
\item{tube}{List containing ETRep details.}
}
\value{
Logical value: TRUE if valid, FALSE otherwise.
}
\description{
Checks the validity of a given ETRep based on the Relative Curvature Condition (RCC) and principal radii such that forall i a_i>b_i.
}
\examples{
# Load tube
data("colon3D")
check_Tube_Legality(tube = colon3D)
}
\references{
Taheri, M., Pizer, S. M., & Schulz, J. (2024). "The Mean Shape under the Relative Curvature Condition." arXiv.
\doi{10.48550/arXiv.2404.01043}

Taheri Shalmani, M. (2024). "Shape Statistics via Skeletal Structures." University of Stavanger.
\doi{10.13140/RG.2.2.34500.23685}
}
