% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETRep_Functions.R
\name{intrinsic_Distance_Between2tubes}
\alias{intrinsic_Distance_Between2tubes}
\title{Calculating the intrinsic distance between two ETReps}
\usage{
intrinsic_Distance_Between2tubes(tube1, tube2)
}
\arguments{
\item{tube1}{List containing ETRep details.}

\item{tube2}{List containing ETRep details.}
}
\value{
Numeric
}
\description{
Calculating the intrinsic distance between two ETReps
}
\examples{
# Load tubes
data("tube_A")
data("tube_B")
intrinsic_Distance_Between2tubes(tube1 = tube_A,tube2 = tube_B)
}
\references{
Taheri, M., Pizer, S. M., & Schulz, J. (2024). "The Mean Shape under the Relative Curvature Condition." arXiv.
\doi{10.48550/arXiv.2404.01043}

Taheri Shalmani, M. (2024). "Shape Statistics via Skeletal Structures." University of Stavanger.
\doi{10.13140/RG.2.2.34500.23685}
}
