% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETRep_Functions.R
\name{nonIntrinsic_Distance_Between2tubes}
\alias{nonIntrinsic_Distance_Between2tubes}
\title{Calculating the non-intrinsic distance between two ETReps}
\usage{
nonIntrinsic_Distance_Between2tubes(tube1, tube2)
}
\arguments{
\item{tube1}{List containing ETRep details.}

\item{tube2}{List containing ETRep details.}
}
\value{
Numeric
}
\description{
Calculating the non-intrinsic distance between two ETReps
}
\examples{
# Load tubes
data("tube_A")
data("tube_B")
intrinsic_Distance_Between2tubes(tube1 = tube_A,tube2 = tube_B)
}
\references{
Taheri, M., Pizer, S. M., & Schulz, J. (2024). "The Mean Shape under the Relative Curvature Condition." arXiv.
\url{https://doi.org/10.48550/arXiv.2404.01043}

Taheri Shalmani, M. (2024). "Shape Statistics via Skeletal Structures." University of Stavanger.
\url{https://uis.brage.unit.no/uis-xmlui/handle/11250/3133161}
}
