% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETRep_Functions.R
\name{plot_Elliptical_Tube}
\alias{plot_Elliptical_Tube}
\title{Plot an Elliptical Tube (ETRep)}
\usage{
plot_Elliptical_Tube(
  e_tube,
  plot_boundary = TRUE,
  plot_r_max = FALSE,
  plot_r_project = TRUE,
  plot_frames = TRUE,
  frameScaling = NA,
  plot_spine = TRUE,
  plot_normal_vec = FALSE,
  plot_skeletal_sheet = TRUE,
  decorate = TRUE,
  colSkeletalSheet = "blue",
  colorBoundary = "blue",
  userMatrix4plotOrientation = NA,
  plotSignificantFeatures = FALSE,
  p_values_v = NULL,
  p_values_psi = NULL,
  p_values_x = NULL,
  p_values_a = NULL,
  p_values_b = NULL,
  lwdSignificant = 4,
  add = FALSE,
  mainTitle = ""
)
}
\arguments{
\item{plot_boundary}{Logical, enables plotting of the boundary (default is TRUE).}

\item{plot_r_max}{Logical, enables plotting of max projection size (default is FALSE).}

\item{plot_r_project}{Logical, enables plotting of projection along normals (default is TRUE).}

\item{plot_frames}{Logical, enables plotting of the material frames (default is TRUE).}

\item{frameScaling}{Numeric, scale factor for frames.}

\item{tube}{List containing ETRep details.}
}
\value{
Graphical output.
}
\description{
Plots a given ETRep with options for boundary, material frames, and projection visualization.
}
\examples{
# Load tube
data("colon3D")
plot_Elliptical_Tube(e_tube = colon3D,
                     plot_frames = FALSE,
                     add=FALSE)
}
